Remove legacy script
